requirements.txt
azure-monitor-opentelemetry==1.6.2
adlfs==2024.7.0
datasketch==1.6.5
duckdb==1.2.0
fastapi==0.115.7
fsspec==2024.10.0
nltk==3.8.1
numpy==2.1.3
opentelemetry-instrumentation-fastapi==0.48b0
pandas==2.2.3
pyarrow==17.0.0
Unidecode==1.3.8
universal_pathlib==0.2.5
uvicorn==0.34.0
xxhash==3.4.1
